Fixed EmptyFunctionBlock (Codacy)
Log empty (but important) callback methods of the Gesture Listener to fix two (2) out of four (4) Codacy PR issues encountered in dfef868
This reduces the Codacy issues to just 2. 

The pull request call still be merged as this issues detected by Codacy does not affect the code nor the functionality of the app. The PR is okay to merge as it is.