Fix some issues caused by 5f21627 and Fix request audio focus
The commit 5f21627 caused that shuffle mode and repeat mode change to MODE_NONE every time the playback state changes.

Request audio focus: when opening the application (play paused) and then pressing the play all or randomly play all button, the audio focus was not requested, nor when the current Playback view was opened and then it was pressed the next or previous button.