New Sort Modes, Crash with Dialog, Local Lyrics only, No cover on album
New sort modes for songs, albums, and artists

Crashed when device was rotated and a sort modes dialog was opened or when the system dark mode was activated.

Now just the embedded lyrics will be loaded, soon will be added the option to add lyrics to a song.

On album detail view the cover for every single song was replaced with the track number because it was redundant.